+++ src/django/api/reports/contributor_first_appearance.sql
@@ -6,7 +6,8 @@ from (
     min(date_part('month', l.created_at)) AS month,
     c.id
   FROM api_facilitylist l
-  JOIN api_contributor c ON l.contributor_id = c.id
+  JOIN api_source s ON s.facility_list_id = l.id
+  JOIN api_contributor c ON s.contributor_id = c.id
   JOIN api_user u ON u.id = c.admin_id
   WHERE date_part('month', l.created_at) != date_part('month', now())
   AND NOT u.email LIKE '%openapparel.org%'