+++ src/django/api/reports/unique_contributor_count_by_type.sql
@@ -8,7 +8,8 @@ FROM (
     c.id,
     c.contrib_type
   FROM api_facilitylist l
-  JOIN api_contributor c ON l.contributor_id = c.id
+  JOIN api_source s ON s.facility_list_id = l.id
+  JOIN api_contributor c ON s.contributor_id = c.id
   JOIN api_user u ON u.id = c.admin_id
   WHERE date_part('month', l.created_at) != date_part('month', now())
   AND NOT u.email like '%openapparel.org%'